Use child_process.execFile instead of .exec for more robust parsing.
In particular, I encountered an error when linting PHP scripts whose
names included a space, because the argument passed to the name wasn't
escaped or quoted.  Rather than using a regex or some other "escaping"
invented on the fly, this commit pushes that responsibility back to the
child_process module.